Rename right.py to rights.py for consistency with route names
