+++ web/views/rights.py
